aac-app/backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose
passlib[bcrypt]
